refactor(webapp-game): move off recruit move to actionMoves
